+++ NEWS.md
@@ -11,7 +11,7 @@
 
 * function `tab_labs()` (with alias `tabl2()`) added: alternative implementation of `tabl()` functionality that always displays tabulations in terms of value labels (for columns that have them), whereas `tabl()` displays tabulations in terms of raw values by default (even for value-labeled variables). `tab_labs()` also incorporates a qtiles argument that temporarily converts any requested non-value-labeled, many-valued numeric variables to quantile category variables, so that they can be included in tabulations.
 
-* function `tabl()` updated with `tab_labs()` -style qtiles argument that allows for non-value-labeled, many-valued numeric variables to be converted temporarily and on the fly to quantile category variables so that they can be included in returned table results. However, this argument is IGNORED UNLESS the user explicitly supplies an integer value between 2 and 100 AND the labs.on argument is set to TRUE.
+* function `tabl()` updated with `tab_labs()` -style qtiles argument that allows for non-value-labeled, many-valued numeric variables to be converted temporarily and on the fly to quantile category variables so that they can be included in returned table results. 
 
 * function `use_val_lab1()` (with alias `uvl1()`) added. This provides the same functionality as `use_val_labs()`, but allows the user to supply only one unquoted variable, whereas `use_val_labs()` requires quoted variable name arguments but allows the user to pass a vector of multiple variable names in one call.